enable Rename Folder

+++ backend/src/router/move.ts
@@ -16,11 +16,23 @@ router.post('/', async (res, resp) => {
     let fileName = res.body["source"].split('/');
     fileName = fileName[fileName.length - 1];
     try {
+        let pointHead = path.join(evn.environment.warehouse,res.body["source"])
+        let pointTail = path.join(evn.environment.warehouse,res.body["destination"])
         let sourceExist = await fs.pathExists(evn.environment.warehouse + "/" + res.body["source"]);
         console.log(evn.environment.warehouse + "/" + res.body["source"]);
         if (sourceExist) {
-            console.log(evn.environment.warehouse + "/" + res.body["source"], evn.environment.warehouse + "/" + res.body["destination"]);
-            await fs.moveSync("./warehouse/" + res.body["source"], "./warehouse/" + res.body["destination"] + "/" + fileName);
+            let isFile = await fs.stat(pointHead);
+            if (isFile.isFile()) {
+                let pointTailThumbNail = path.join(pointTail,'.thumbnail')
+                let thumbnailName = path.basename(pointHead)+".png"
+                let pathToThumbNail = path.join(path.dirname(pointHead),".thumbnail",thumbnailName)
+                fs.moveSync(pathToThumbNail,path.join(pointTailThumbNail,thumbnailName))
+                await fs.moveSync("./warehouse/" + res.body["source"], "./warehouse/" + res.body["destination"] + "/" + fileName);
+            }else{
+                console.log(evn.environment.warehouse + "/" + res.body["source"], evn.environment.warehouse + "/" + res.body["destination"]);
+                await fs.moveSync("./warehouse/" + res.body["source"], "./warehouse/" + res.body["destination"] + "/" + fileName);
+            }
+
             resp.send({result:"Folder/file " + res.body["destination"] + " is move"});
         } else {
             resp.send({result:'Folder/file is not exist !!!'});
